+++ CHANGELOG.md
@@ -56,6 +56,7 @@ required).
 ### Fixed
 - tc-72 : The DestinationsAdmin page could only display 200 first destinations
 - tc-150 : Admin could not display non concurrency rules for commercial lines in subfolders
+- In RoutePlannerFunction and PTJourneyPlannerService : max depth (number of journeys) is number of connections + 1
 
 ### Security